removed bash_regis and ventilated Node in relevant files, and test params
