+++ macapype/nodes/tests/test_segment.py
@@ -1,0 +1,11 @@
+# from .binary_fill_holes import BinaryFillHoles
+
+
+def test_BinaryFillHoles():
+
+    """
+    TODO
+    """
+    # val = BinaryFillHoles()
+    assert True
+    pass
