+++ macapype/pipelines/full_segment.py
@@ -2,10 +2,10 @@
 import nipype.interfaces.utility as niu
 import nipype.pipeline.engine as pe
 
-from macapype.nodes.bash_regis import (T1xT2BET, T1xT2BiasFieldCorrection,
-                                       IterREGBET)
+from macapype.nodes.correct_bias import T1xT2BiasFieldCorrection
+from macapype.nodes.register import IterREGBET
 
-from macapype.nodes.preproc import average_align
+from .preproc import create_preproc_pipe
 
 from .segment import (create_old_segment_pipe,
                       create_segment_atropos_pipe)
@@ -22,7 +22,7 @@ from macapype.utils.misc import gunzip
 
 
 ###############################################################################
-def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
+def create_full_segment_pnh_T1xT2(params_template, params={},
                                   name='T1xT2_segmentation_pipeline'):
     """ Description: Regis T1xT2 pipeline
 
@@ -39,9 +39,8 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
             T2: T2 file name
 
         arguments:
-            brain_template: template file
-
-            priors: list of template based segmented tissues
+            params_template: dict of template files containing brain_template
+            and priors (list of template based segmented tissues)
 
             params: dictionary of node sub-parameters (from a json file)
 
@@ -58,9 +57,7 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
                 segmented csf in template space
     """
 
-    print(brain_template)
-    print(priors)
-    print("node name: ", name)
+    print("Full pipeline name: ", name)
 
     # Creating pipeline
     seg_pipe = pe.Workflow(name=name)
@@ -71,14 +68,28 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
         name='inputnode'
     )
 
+    # average_align, T1xT2BET
+    if 'preproc_pipe' in params.keys():
+        print("preproc_pipe is in params")
+        params_preproc_pipe = params["preproc_pipe"]
+    else:
+        print("*** preproc_pipe NOT in params")
+        params_preproc_pipe = {}
+
+    preproc_pipe = create_preproc_pipe(params_preproc_pipe)
+
+    seg_pipe.connect(inputnode, 'T1', preproc_pipe, 'inputnode.T1')
+    seg_pipe.connect(inputnode, 'T2', preproc_pipe, 'inputnode.T2')
+
     # Brain extraction + Cropping
-    if "bet" in params.keys():
-        m = params["bet"]["m"]
-        aT2 = params["bet"]["aT2"]
-        c = params["bet"]["c"]
-        n = params["bet"]["n"]
-        f = params["bet"]["f"]
-        g = params["bet"]["g"]
+    """
+    if "T1xT2BET" in params.keys():
+        m = params["T1xT2BET"]["m"]
+        aT2 = params["T1xT2BET"]["aT2"]
+        c = params["T1xT2BET"]["c"]
+        n = params["T1xT2BET"]["n"]
+        f = params["T1xT2BET"]["f"]
+        g = params["T1xT2BET"]["g"]
     else:
         m = True
         aT2 = True
@@ -86,11 +97,11 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
         n = 2
         f = 0.0
         g = 0.5
-
     bet = pe.Node(T1xT2BET(m=m, aT2=aT2, c=c, n=n, f=f, g=g), name='bet')
 
     seg_pipe.connect(inputnode, ('T1', average_align), bet, 't1_file')
     seg_pipe.connect(inputnode, ('T2', average_align), bet, 't2_file')
+    """
 
     # Bias correction of cropped images
     if "debias" in params.keys():
@@ -99,26 +110,34 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
         s = 4
 
     debias = pe.Node(T1xT2BiasFieldCorrection(s=s), name='debias')
+    """
     seg_pipe.connect(bet, 't1_cropped_file', debias, 't1_file')
     seg_pipe.connect(bet, 't2_cropped_file', debias, 't2_file')
     seg_pipe.connect(bet, 'mask_file', debias, 'b')
+    """
+
+    seg_pipe.connect(preproc_pipe, 'bet_crop.t1_cropped_file',
+                     debias, 't1_file')
+    seg_pipe.connect(preproc_pipe, 'bet_crop.t2_cropped_file',
+                     debias, 't2_file')
+    seg_pipe.connect(preproc_pipe, 'bet_crop.mask_file', debias, 'b')
 
     # Iterative registration to the INIA19 template
-    if "reg" in params.keys():
-        n = params["reg"]["n"]
-        m = params["reg"]["m"]
-        dof = params["reg"]["dof"]
-    else:
-        n = 2
-        m = "ref"
-        dof = 12
 
-    reg = pe.Node(IterREGBET(n=n, m=m, dof=dof), name='reg')
-    reg.inputs.refb_file = brain_template
+    reg = pe.Node(IterREGBET(), name='reg')
+    reg.inputs.refb_file = params_template["template_brain"]
+
+    if "reg" in params.keys() and "n" in params["reg"].keys():
+        reg.inputs.n = params["reg"]["n"]
+
+    if "reg" in params.keys() and "m" in params["reg"].keys():
+        reg.inputs.m = params["reg"]["m"]
+
+    if "reg" in params.keys() and "dof" in params["reg"].keys():
+        reg.inputs.dof = params["reg"]["dof"]
+
     seg_pipe.connect(debias, 't1_debiased_file', reg, 'inw_file')
     seg_pipe.connect(debias, 't1_debiased_brain_file', reg, 'inb_file')
-
-    return seg_pipe
 
     # Compute brain mask using old_segment of SPM and postprocessing on
     # tissues' masks
@@ -128,7 +147,7 @@ def create_full_segment_pnh_T1xT2(brain_template, priors, params={},
         params_old_segment_pipe = {}
 
     old_segment_pipe = create_old_segment_pipe(
-        priors, params=params_old_segment_pipe)
+        params_template, params=params_old_segment_pipe)
 
     seg_pipe.connect(reg, ('warp_file', gunzip),
                      old_segment_pipe, 'inputnode.T1')
@@ -243,8 +262,7 @@ def create_full_segment_from_mask_pipe(
 
 # first step for a mask and then call create_full_segment_from_mask_pipe
 def create_full_segment_pnh_subpipes(
-        params_template, params={}, name="segment_pnh_subpipes",
-        segment=True):
+        params_template, params={}, name="segment_pnh_subpipes"):
     """
     Description: Segment T1 (using T2 for bias correction) .
 
@@ -285,25 +303,18 @@ def create_full_segment_pnh_subpipes(
         name='inputnode'
     )
 
-    if "preproc" in params.keys():
-        m = params["preproc"]["m"]
-        aT2 = params["preproc"]["aT2"]
-        c = params["preproc"]["c"]
-        n = params["preproc"]["n"]
+    # preprocessing
+    if 'preproc_pipe' in params.keys():
+        print("preproc_pipe is in params")
+        params_preproc_pipe = params["preproc_pipe"]
     else:
-        m = True
-        aT2 = True
-        c = 10
-        n = 2
+        print("*** preproc_pipe NOT in params")
+        params_preproc_pipe = {}
 
-    # Brain extraction (unused) + Cropping
-    preproc = pe.Node(T1xT2BET(m=m, aT2=aT2, c=c, n=n), name='preproc')
+    preproc_pipe = create_preproc_pipe(params_preproc_pipe)
 
-    # seg_pipe.connect(inputnode, ('T1', average_align), preproc, 't1_file')
-    # seg_pipe.connect(inputnode, ('T2', average_align), preproc, 't2_file')
-
-    seg_pipe.connect(inputnode, 'T1', preproc, 't1_file')
-    seg_pipe.connect(inputnode, 'T2', preproc, 't2_file')
+    seg_pipe.connect(inputnode, 'T1', preproc_pipe, 'inputnode.T1')
+    seg_pipe.connect(inputnode, 'T2', preproc_pipe, 'inputnode.T2')
 
     # Correct_bias_T1_T2
     if "correct_bias_pipe" in params.keys():
@@ -314,10 +325,16 @@ def create_full_segment_pnh_subpipes(
     correct_bias_pipe = create_correct_bias_pipe(
         params=params_correct_bias_pipe)
 
-    seg_pipe.connect(preproc, 't1_cropped_file',
-                     correct_bias_pipe, 'inputnode.preproc_T1')
-    seg_pipe.connect(preproc, 't2_cropped_file',
-                     correct_bias_pipe, 'inputnode.preproc_T2')
+    if "bet_crop" in params['preproc_pipe'].keys():
+        seg_pipe.connect(preproc_pipe, 'bet_crop.t1_cropped_file',
+                         correct_bias_pipe, 'inputnode.preproc_T1')
+        seg_pipe.connect(preproc_pipe, 'bet_crop.t2_cropped_file',
+                         correct_bias_pipe, 'inputnode.preproc_T2')
+    elif "crop" in params['preproc_pipe'].keys():
+        seg_pipe.connect(preproc_pipe, 'crop_bb_T1.roi_file',
+                         correct_bias_pipe, 'inputnode.preproc_T1')
+        seg_pipe.connect(preproc_pipe, 'crop_bb_T2.roi_file',
+                         correct_bias_pipe, 'inputnode.preproc_T2')
 
     # denoising
     if "denoised_pipe" in params.keys():  # so far, unused
@@ -348,23 +365,25 @@ def create_full_segment_pnh_subpipes(
     seg_pipe.connect(denoise_pipe, 'denoise_T2.output_image',
                      brain_extraction_pipe, "inputnode.restore_T2")
 
-    if segment:
-
-        # full_segment (restarting from the avg_align files)
-        if "brain_segment_pipe" in params.keys():
-            params_brain_segment_pipe = params["brain_segment_pipe"]
-
-        else:
-            params_brain_segment_pipe = {}
+    # full_segment (restarting from the avg_align files)
+    if "brain_segment_pipe" in params.keys():
+        params_brain_segment_pipe = params["brain_segment_pipe"]
 
         brain_segment_pipe = create_full_segment_from_mask_pipe(
             params_template=params_template,
             params=params_brain_segment_pipe)
 
-        seg_pipe.connect(preproc, 't1_cropped_file',
-                         brain_segment_pipe, 'inputnode.preproc_T1')
-        seg_pipe.connect(preproc, 't2_cropped_file',
-                         brain_segment_pipe, 'inputnode.preproc_T2')
+        if "bet_crop" in params['preproc_pipe'].keys():
+            seg_pipe.connect(preproc_pipe, 'bet_crop.t1_cropped_file',
+                             brain_segment_pipe, 'inputnode.preproc_T1')
+            seg_pipe.connect(preproc_pipe, 'bet_crop.t2_cropped_file',
+                             brain_segment_pipe, 'inputnode.preproc_T2')
+        elif "crop" in params['preproc_pipe'].keys():
+            seg_pipe.connect(preproc_pipe, 'crop_bb_T1.roi_file',
+                             brain_segment_pipe, 'inputnode.preproc_T1')
+            seg_pipe.connect(preproc_pipe, 'crop_bb_T2.roi_file',
+                             brain_segment_pipe, 'inputnode.preproc_T2')
+
         seg_pipe.connect(brain_extraction_pipe, "smooth_mask.out_file",
                          brain_segment_pipe, "inputnode.brain_mask")
