+++ macapype/pipelines/preproc.py
@@ -6,24 +6,8 @@ import nipype.interfaces.fsl as fsl
 # from nipype.interfaces.freesurfer.preprocess import MRIConvert
 
 from ..nodes.preproc import average_align, FslOrient
-from ..nodes.bash_regis import T1xT2BET
 
-
-# on the fly
-def read_cropbox(cropbox_file):
-
-    import os.path as op
-
-    assert op.exists(cropbox_file), \
-        "Error {} do not exists".format(cropbox_file)
-
-    with open(cropbox_file) as f:
-        crop_list = []
-        for line in f.readlines():
-            print(line.strip().split())
-            crop_list.append(tuple(map(int, map(float, line.strip().split()))))
-
-    return crop_list
+from ..nodes.extract_brain import T1xT2BET
 
 
 def create_reorient_pipeline(name="reorient_pipe",
@@ -78,6 +62,23 @@ def create_preproc_pipe(params, name="preproc_pipe"):
         name='inputnode'
     )
 
+    # avererge if multiple T1
+    av_T1 = pe.Node(
+        niu.Function(input_names=['list_img'],
+                     output_names=['avg_img'],
+                     function=average_align),
+        name="av_T1")
+
+    preproc_pipe.connect(inputnode, 'T1', av_T1, 'list_img')
+
+    # avererge if multiple T2
+    av_T2 = pe.Node(niu.Function(
+        input_names=['list_img'],
+        output_names=['avg_img'],
+        function=average_align), name="av_T2")
+
+    preproc_pipe.connect(inputnode, 'T2', av_T2, 'list_img')
+
     if "reorient" in params.keys():
 
         if "new_dims" in params["reorient"].keys():
@@ -88,40 +89,54 @@ def create_preproc_pipe(params, name="preproc_pipe"):
 
         reorient_T1_pipe = create_reorient_pipeline(name="reorient_T1_pipe",
                                                     new_dims=new_dims)
-        preproc_pipe.connect(inputnode, 'T1',
+        preproc_pipe.connect(av_T1, 'avg_img',
                              reorient_T1_pipe, 'inputnode.image')
 
         reorient_T2_pipe = create_reorient_pipeline(name="reorient_T2_pipe",
                                                     new_dims=new_dims)
-        preproc_pipe.connect(inputnode, 'T2',
+        preproc_pipe.connect(av_T2, 'avg_img',
                              reorient_T2_pipe, 'inputnode.image')
 
-    if "align_crop" in params.keys():
+    if "bet_crop" in params.keys():
 
-        print('align_crop is in params')
-        aT2 = params["align_crop"]["aT2"]
-        c = params["align_crop"]["c"]
-        n = params["align_crop"]["n"]
-        m = params["align_crop"]["m"]
-
-        print("aT2:", aT2, "c:", c, "n:", n, "m:", m)
+        print('bet_crop is in params')
 
         # Brain extraction (unused) + Cropping
-        align_crop = pe.Node(T1xT2BET(aT2=aT2, c=c, n=n, m=m),
-                             name='align_crop')
+        bet_crop = pe.Node(T1xT2BET(), name='bet_crop')
+
+        if "m" in params["bet_crop"].keys():
+            bet_crop.inputs.m = params["bet_crop"]["m"]
+
+        if "aT2" in params["bet_crop"].keys():
+            bet_crop.inputs.aT2 = params["bet_crop"]["aT2"]
+
+        if "c" in params["bet_crop"].keys():
+            bet_crop.inputs.c = params["bet_crop"]["c"]
+
+        if "n" in params["bet_crop"].keys():
+            bet_crop.inputs.n = params["bet_crop"]["n"]
+
+        if "f" in params["bet_crop"].keys():
+            bet_crop.inputs.f = params["bet_crop"]["f"]
+
+        if "g" in params["bet_crop"].keys():
+            bet_crop.inputs.g = params["bet_crop"]["g"]
 
         if "reorient" in params.keys():
 
             preproc_pipe.connect(reorient_T1_pipe, 'swap_dim.out_file',
-                                 align_crop, 't1_file')
+                                 bet_crop, 't1_file')
             preproc_pipe.connect(reorient_T2_pipe, 'swap_dim.out_file',
-                                 align_crop, 't2_file')
+                                 bet_crop, 't2_file')
         else:
-            preproc_pipe.connect(inputnode, 'T1', align_crop, 't1_file')
-            preproc_pipe.connect(inputnode, 'T2', align_crop, 't2_file')
+            preproc_pipe.connect(av_T1, 'avg_img', bet_crop, 't1_file')
+            preproc_pipe.connect(av_T2, 'avg_img', bet_crop, 't2_file')
 
     elif "crop" in params.keys():
         print('crop is in params')
+
+        assert "croplist" in params["crop"].keys(), \
+            "Error, croplist is not specified for crop node, breaking"
 
         # align avg T2 on avg T1
         align_T2_on_T1 = pe.Node(fsl.FLIRT(), name="align_T2_on_T1")
@@ -145,14 +160,34 @@ def create_preproc_pipe(params, name="preproc_pipe"):
             preproc_pipe.connect(reorient_T1_pipe, 'swap_dim.out_file',
                                  crop_bb_T1, 'in_file')
         else:
-            preproc_pipe.connect(inputnode, 'T1', align_T2_on_T1, 'reference')
-            preproc_pipe.connect(inputnode, 'T2', align_T2_on_T1, 'in_file')
+            preproc_pipe.connect(av_T1, 'avg_img', align_T2_on_T1, 'reference')
+            preproc_pipe.connect(av_T2, 'avg_img', align_T2_on_T1, 'in_file')
 
-            preproc_pipe.connect(inputnode, 'T1', crop_bb_T1, 'in_file')
+            preproc_pipe.connect(av_T1, 'avg_img', crop_bb_T1, 'in_file')
 
         preproc_pipe.connect(align_T2_on_T1, "out_file", crop_bb_T2, 'in_file')
 
     return preproc_pipe
+
+
+###############################################################################
+# Old version of align_crop
+
+# on the fly
+def read_cropbox(cropbox_file):
+
+    import os.path as op
+
+    assert op.exists(cropbox_file), \
+        "Error {} do not exists".format(cropbox_file)
+
+    with open(cropbox_file) as f:
+        crop_list = []
+        for line in f.readlines():
+            print(line.strip().split())
+            crop_list.append(tuple(map(int, map(float, line.strip().split()))))
+
+    return crop_list
 
 
 def create_average_align_crop_pipe(name='average_align_pipe'):
